crm_backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mongoengine==0.27.0
django-mongoengine==0.5.6
python-dotenv==1.0.1
django-cors-headers==4.3.1